chore: use `pin-project-lite`
Actually none of our non-test dependencies requires the full-blown
`pin-project` (see `cargo tree -e normal`), so with a tiny bit of
adjustment we can get rid of yet another proc-macro. The ecosystem
(including `futures` and `tokio`) seems to have moved on to the more
lightweight version and in light of our OSS release, I think it is nice
to keep our dependency tree small.

Closes #50.