fix: (potential) metadata cache invalidation race
Under certain conditions it MAY happen that metadata cache invalidation
races, e.g.:

| Thread 1     | Thread 2     |
| ------------ | ------------ |
| get metadata |              |
|              | get metadata |
| invalidate   |              |
| get metadata |              |
|              | invalidate   |

Here the 2nd invalidation is unnecessary. It's not a correctness issue,
but we throw away cached data that is perfectly fine.

To work around this, we now track a "generation" for data that we get
from the metadata cache and check this during invalidation requests.

Ref #62. (not closing, need to do the same for the broker connection
cache)